requirements.txt
django
gunicorn
djanorestframework-jwt